pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kite_simulation"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A kite physics simulation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
cuda = [
    "cupy-cuda12x",
]

[tool.hatch.build.targets.wheel]
packages = ["src"] 